requirements.txt
Django>=4.0,<5.0
psycopg2-binary<3.0,>=2.9
requests
pyotp
Flask
telebot
drf-yasg
djangorestframework-simplejwt
Pillow
django-simple-captcha~=0.6.2